Update owners.js
Changed "Tim Chien" to "Tim Guan-tin Chien" so it is consistent with phonebook. The release management nag tool is utilizing this js file to retrieve component owners and send out any reminders that may be necessary. The tool relies on using the python phonebook module to get the email addresses, hence needed the right names.